Use storage.Watch OnCapabilityUpdate/OnStateUpdate/Fingerprint callbacks
Update startAlexaWatch to supply an alexaFingerprint function via the new
Fingerprint hook and route updates through OnCapabilityUpdate vs OnStateUpdate.
Clean up capFingerprint → alexaFingerprint rename.

+++ cmd/plugin-alexa/test.bdd.steps_test.go
@@ -573,7 +573,7 @@ func (c *bddCtx) lightTemperatureIs(expected int) error {
 	if !ok {
 		return fmt.Errorf("state type: got %T, want domain.Light", c.lastEntity.State)
 	}
-	if st.Temperature != float64(expected) {
+	if st.Temperature != expected {
 		return fmt.Errorf("light.Temperature: got %v, want %v", st.Temperature, expected)
 	}
 	return nil
